fix:Move validation from link to mandatory validator [DHIS2-14711][2.40]
