+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/validation/validator/relationship/LinkValidator.java
@@ -28,16 +28,12 @@
 package org.hisp.dhis.tracker.validation.validator.relationship;
 
 import static org.hisp.dhis.tracker.validation.ValidationCode.E4000;
-import static org.hisp.dhis.tracker.validation.ValidationCode.E4001;
 
 import java.util.Objects;
-import java.util.stream.Stream;
 
-import org.apache.commons.lang3.StringUtils;
 import org.hisp.dhis.tracker.TrackerImportStrategy;
 import org.hisp.dhis.tracker.bundle.TrackerBundle;
 import org.hisp.dhis.tracker.domain.Relationship;
-import org.hisp.dhis.tracker.domain.RelationshipItem;
 import org.hisp.dhis.tracker.validation.Reporter;
 import org.hisp.dhis.tracker.validation.Validator;
 
@@ -46,17 +42,7 @@ public class LinkValidator implements Validator<Relationship>
     @Override
     public void validate( Reporter reporter, TrackerBundle bundle, Relationship relationship )
     {
-        validateRelationshipItemsContainOnlyOneEntity( reporter, relationship );
         validateRelationshipDoesNotLinkEntityToItself( reporter, relationship );
-    }
-
-    private void validateRelationshipItemsContainOnlyOneEntity( Reporter reporter,
-        Relationship relationship )
-    {
-        reporter.addErrorIf( () -> hasUnexpectedReferences( relationship.getFrom() ),
-            relationship, E4001, "from", relationship.getRelationship() );
-        reporter.addErrorIf( () -> hasUnexpectedReferences( relationship.getTo() ),
-            relationship, E4001, "to", relationship.getRelationship() );
     }
 
     private void validateRelationshipDoesNotLinkEntityToItself( Reporter reporter, Relationship relationship )
@@ -65,13 +51,6 @@ public class LinkValidator implements Validator<Relationship>
         {
             reporter.addError( relationship, E4000, relationship.getRelationship() );
         }
-    }
-
-    private boolean hasUnexpectedReferences( RelationshipItem item )
-    {
-        return Stream.of( item.getTrackedEntity(), item.getEnrollment(), item.getEvent() )
-            .filter( StringUtils::isNotBlank )
-            .count() != 1;
     }
 
     @Override
